refactor: migrate contest controller to TypeScript
